Restructure files into namespaces and subdirectories
