+++ src/CMakeLists.txt
@@ -17,7 +17,7 @@ add_subdirectory(query/v2)
 add_subdirectory(slk)
 add_subdirectory(rpc)
 add_subdirectory(auth)
-add_subdirectory(io/v3)
+add_subdirectory(io/simulator)
 
 if (MG_ENTERPRISE)
   add_subdirectory(audit)
